GeecsBluesky 0.77.1: extract the scan preamble to plans/preamble.py (#807 phase 2 step A)
A pure move, no behaviour change: resolve_request (validate + trigger
profile -> ShotController + strict) and prepare_step_scan (save sets,
unserved/CONNECTED preflights, action slots, axis resolution, capture
toggle, deferred construction, in-plan connects, fail-fast gates, the md
spec) leave the funnel so that one implementation has two callers — the
funnel today, the RunEngine preprocessor next. The funnel keeps working
unchanged and loses ~300 lines; PreparedScan.as_claimed_kwargs() keeps its
build_claimed_scan_plan call a one-liner. _DeferredConnectFactories, the
connect/disconnect stubs and their two constants moved with the code.

Tests: 47 test_scan_request_plan (document parity + every pre-claim
invariant), 127 across runner/wrapper/named-plans/startup/optimize/action,
50 across pause/abort/staging/grid/native-save/s-file/free-run — all green.
The four patch targets for make_scalar_policy / ShotController moved to
geecs_bluesky.plans.preamble.
