+++ GeecsBluesky/geecs_bluesky/plans/scan_request_plan.py
@@ -52,7 +52,6 @@ worker-wide default session via :func:`set_plan_session`) later.
 
 from __future__ import annotations
 
-import asyncio
 import json
 import logging
 import math
@@ -62,15 +61,21 @@ from typing import Any, Callable
 import bluesky.plan_stubs as bps
 import bluesky.preprocessors as bpp
 from bluesky.utils import RequestAbort
-from ophyd_async.plan_stubs import ensure_connected
 
 from geecs_bluesky.config_resolver import ConfigResolver, ConfigsRepoResolver
-from geecs_bluesky.db_runtime import select_telemetry_variables
 from geecs_bluesky.exceptions import GeecsConfigurationError
 from geecs_bluesky.optimize import BinData
 from geecs_bluesky.plans.optimize import geecs_adaptive_scan
 from geecs_bluesky.plans.orchestration import _with_pause_quiescer
 from geecs_bluesky.plans.pause_semantics import ShotControlPauseQuiescer
+from geecs_bluesky.plans.preamble import (
+    _await_in_plan,
+    _connect_in_batches,
+    _DeferredConnectFactories,
+    _disconnect_plan,
+    prepare_step_scan,
+    resolve_request,
+)
 from geecs_bluesky.plans.run_wrapper import (
     claim_scan,
     claim_scan_number,
@@ -85,30 +90,24 @@ from geecs_bluesky.scan_log import (
 from geecs_bluesky.scan_request_runner import (
     PseudoMovableTarget,
     _build_request_detectors,
-    _defaults_flag,
     _preflight_connected,
     _preflight_unserved,
-    assemble_action_slots,
     build_action_registry,
     build_movable,
-    build_step_scan_spec,
     compile_action_slot,
     make_scalar_policy,
     metadata_applied_defaults,
     metadata_submission,
     merge_optimizer_device_requirements,
     prefetch_action_signals,
-    resolve_and_apply_capture_toggle,
     resolve_movable_target,
     resolve_save_sets_and_rituals,
     save_set_to_devices_config,
-    trigger_writes_from_profile,
-    validate_scan_request,
     warn_if_reserved_boundary_overrides,
 )
 from geecs_bluesky.session import _json_safe
 from geecs_bluesky.shot_controller import ShotController
-from geecs_schemas import AcquisitionMode, ScanRequest, ScanRequestMode
+from geecs_schemas import ScanRequest, ScanRequestMode
 
 logger = logging.getLogger(__name__)
 
@@ -202,28 +201,6 @@ RUN_ACTION_PLAN_ANNOTATION: dict = {
     },
 }
 
-#: Per-connect timeout for the in-plan strict batches.  Near-parity, not
-#: parity, with ``GeecsSession._connect``: there 20 s is the outer wall on
-#: the loop hop while each connect runs at the ophyd-async default (10 s),
-#: so an unreachable device fails at ~10 s; here the 20 s *is* the
-#: per-connect timeout, so the same failure surfaces ~10 s later.
-_CONNECT_TIMEOUT = 20.0
-
-#: The session factory methods whose *construction* code the plan reuses
-#: verbatim (only the connect step is deferred — see
-#: :class:`_DeferredConnectFactories`).
-_SESSION_FACTORIES = frozenset(
-    {
-        "detector",
-        "contributor",
-        "snapshot",
-        "motor",
-        "settable",
-        "confirm_settable",
-        "pseudo_movable",
-        "action_signal_factory",
-    }
-)
 
 #: The worker-wide default session (installed once at worker startup).
 _worker_session: Any | None = None
@@ -278,155 +255,6 @@ def set_optimization_loader(loader: Callable[[Any], Any] | None) -> None:
     """
     global _optimization_loader
     _optimization_loader = loader
-
-
-class _DeferredConnectFactories:
-    """A session facade: identical device construction, connect deferred.
-
-    The session's factory methods (``detector``, ``motor``, …) are the one
-    definition of how request devices are constructed — but each ends in
-    ``self._connect(device)``, a blocking hop onto the RE loop that would
-    deadlock inside a plan.  This facade binds those *same* class functions
-    to itself (so the construction code cannot drift from the session's)
-    and swaps only ``_connect``: devices are recorded on :attr:`created`
-    and connected later in one in-plan batch.  Everything else (attributes
-    like ``experiment`` / ``_mock`` / ``rep_rate_hz``) delegates to the
-    wrapped session.
-
-    Deliberately *not* covering ``telemetry``/``telemetry_batch``: the soft
-    tier's connect-failure-drops semantics need their own in-plan gather
-    (:func:`_connect_telemetry_plan`), not the strict batch connect.
-    """
-
-    def __init__(self, session: Any) -> None:
-        self._session = session
-        self.created: list = []
-
-    def _connect(self, device: Any) -> Any:
-        self.created.append(device)
-        return device
-
-    def __getattr__(self, name: str) -> Any:
-        if name in _SESSION_FACTORIES:
-            func = getattr(type(self._session), name, None)
-            if func is not None:
-                return func.__get__(self, type(self))
-        return getattr(self._session, name)
-
-
-def _await_in_plan(coro_fn: Any):
-    """Plan stub: await one no-arg coroutine function, propagating its error.
-
-    ``bps.wait_for`` alone parks the plan on the future but discards its
-    outcome; re-raising through ``task.result()`` keeps in-plan connects
-    fail-fast (the ophyd-async ``wait_for_awaitable`` idiom, not exported
-    by the pinned release).
-    """
-    tasks = yield from bps.wait_for([coro_fn])
-    return tasks[0].result()
-
-
-def _connect_in_batches(devices: list, *, mock: bool):
-    """Plan stub: strict-connect *devices* via ``ensure_connected``.
-
-    ``ensure_connected`` refuses duplicate device names within one call, and
-    an action-check signal can legitimately share a name with a scan-axis
-    movable on the same ``Device:Variable`` — so the list is split into
-    unique-name batches instead of failing the scan on a naming accident.
-    Failures propagate (strict tier fails loudly), pre-claim.
-    """
-    pending = list(devices)
-    while pending:
-        batch: list = []
-        rest: list = []
-        seen: set[str] = set()
-        for device in pending:
-            if device.name in seen:
-                rest.append(device)
-            else:
-                seen.add(device.name)
-                batch.append(device)
-        yield from ensure_connected(*batch, mock=mock, timeout=_CONNECT_TIMEOUT)
-        pending = rest
-
-
-def _connect_telemetry_plan(session: Any, save_set: Any, scalar_policy: Any):
-    """Plan stub: build + soft-connect the Tier-2 telemetry group in-plan.
-
-    Same selection (:func:`~geecs_bluesky.db_runtime.select_telemetry_variables`),
-    same soft-tier contract (a device unreachable at scan start is dropped
-    with a warning, never an abort; only connected devices are recorded),
-    one :class:`~geecs_bluesky.devices.ca.telemetry.CaTelemetryGroup` per
-    scan.  One concurrent gather, so wall time is the slowest device.
-
-    Returns
-    -------
-    tuple
-        ``(readables, recorded)`` — the group (or nothing) to append to the
-        read set, and ``{device: [variables]}`` of what actually connected.
-    """
-    if scalar_policy is None:
-        return [], {}
-    selected = select_telemetry_variables(
-        save_set, scalar_policy.subscribed_by_device()
-    )
-    if not selected:
-        return [], {}
-    # Lazy import: keeps this module importable without the `ca` extra
-    # (same discipline as the runner).
-    from geecs_bluesky.devices.ca.telemetry import CaTelemetryGroup, CaTelemetryReadable
-
-    members = [
-        CaTelemetryReadable(device, variables, experiment=session.experiment)
-        for device, variables in selected.items()
-    ]
-    results: list = []
-
-    async def _connect_all() -> None:
-        results.extend(
-            await asyncio.gather(
-                *(m.connect(mock=session._mock) for m in members),
-                return_exceptions=True,
-            )
-        )
-
-    yield from bps.wait_for([_connect_all])
-
-    connected: list = []
-    recorded: dict[str, list[str]] = {}
-    for member, variables, result in zip(members, selected.values(), results):
-        if isinstance(result, BaseException):
-            logger.warning(
-                "Dropping background-telemetry device %s: unreachable at scan "
-                "start (soft tier — never aborts the scan)",
-                member._geecs_device_name,
-                exc_info=result,
-            )
-        else:
-            connected.append(member)
-            recorded[member._geecs_device_name] = list(variables)
-    if not connected:
-        return [], recorded
-    return [CaTelemetryGroup(connected)], recorded
-
-
-def _disconnect_plan(created: list):
-    """Plan stub: best-effort disconnect of everything the plan created.
-
-    The in-plan counterpart of the runner's ``finally:
-    session.disconnect(*created)`` (which hops onto the RE loop).  Runs as
-    a finalize, so it executes on success, failure, and abort alike; each
-    device's failure is swallowed (gather with exceptions returned) —
-    cleanup never masks the plan's own outcome.
-    """
-    closers = [d for d in created if hasattr(d, "disconnect")]
-    if not closers:
-        return
-
-    async def _disconnect_all() -> None:
-        await asyncio.gather(*(d.disconnect() for d in closers), return_exceptions=True)
-
-    yield from bps.wait_for([_disconnect_all])
 
 
 # Signature constraint (queueserver contract): at `queue add` the RE Manager
@@ -593,23 +421,10 @@ def _scan_request_body(
     *failed_move_policy* are the two front-door seams (see the plan's doc).
     """
     # ---- phase 1: authoritative validation + pure resolution (pre-claim) --
-    if not isinstance(request, ScanRequest):
-        request = ScanRequest.model_validate(request)
-    request, applied_defaults, defaults = validate_scan_request(request, resolver)
-
-    controller = None
-    if request.capture.trigger_profile:
-        profile = resolver.resolve_trigger_profile(request.capture.trigger_profile)
-        writes = trigger_writes_from_profile(profile)
-        if writes.states:
-            # Constructed worker-side, unconnected; the setter reachability
-            # check joins the in-plan connect stage below.
-            controller = ShotController.from_writes(
-                writes,
-                experiment=session.experiment,
-                rep_rate_hz=session.rep_rate_hz,
-            )
-    strict = request.capture.acquisition is AcquisitionMode.STRICT
+    resolved = resolve_request(session, resolver, request)
+    request = resolved.request
+    controller = resolved.controller
+    strict = resolved.strict
 
     if request.mode is ScanRequestMode.OPTIMIZE:
         if request.capture.native_image_save is False:
@@ -628,124 +443,16 @@ def _scan_request_body(
                 controller,
                 strict,
                 created,
-                applied_defaults=applied_defaults,
+                applied_defaults=resolved.applied_defaults,
                 submission=submission,
                 optimization_loader=optimization_loader,
             )
         )
 
-    save_set, rituals = resolve_save_sets_and_rituals(
-        resolver, request.capture.save_sets
+    # ---- phases 1b-3: the shared pre-claim preamble (plans/preamble.py) ----
+    prepared = yield from prepare_step_scan(
+        session, resolver, resolved, created, submission=submission
     )
-    scalar_policy = make_scalar_policy(session)
-    devices_config = save_set_to_devices_config(save_set, scalar_policy)
-    # Unserved-variables check, headless by decision 3 (operator questions
-    # are client-side pre-submit): continue-and-drop with a WARNING.
-    checked_config, dropped_unserved, dropped_unserved_devices = _preflight_unserved(
-        session, devices_config
-    )
-    if checked_config is None:  # defensive: the headless default never aborts
-        raise GeecsConfigurationError(
-            "unserved-variables pre-flight aborted the scan (pre-claim)"
-        )
-    devices_config = checked_config
-    # CONNECTED liveness re-check (#664): the client asked pre-submit, but
-    # the queue's submission-to-execution gap is long — re-check here,
-    # refusing only when a row could never complete.
-    disconnected_devices = _preflight_connected(session, devices_config)
-    slots = assemble_action_slots(request.actions, applied_defaults, rituals)
-    warn_if_reserved_boundary_overrides(save_set)
-    axis_resolved = [
-        resolve_movable_target(
-            resolver.resolve_scan_variable(axis.variable), axis.variable
-        )
-        for axis in request.axes
-    ]
-    telemetry_enabled = (
-        request.capture.background_telemetry
-        if request.capture.background_telemetry is not None
-        else _defaults_flag(defaults, "background_telemetry", True)
-    )
-    devices_config, capture_devices, native_image_save = (
-        resolve_and_apply_capture_toggle(request, defaults, devices_config, session)
-    )
-
-    # ---- phase 2: worker-side construction (connects deferred) -----------
-    factories = _DeferredConnectFactories(session)
-    setup = per_step = closeout = None
-    if any(slots.values()):
-        factory = factories.action_signal_factory()
-        created.append(factory)
-        registry = build_action_registry(resolver)
-        setup, setup_plans = compile_action_slot(
-            slots["setup"], resolver, registry, factory
-        )
-        per_step, per_step_plans = compile_action_slot(
-            slots["per_step"], resolver, registry, factory
-        )
-        closeout, closeout_plans = compile_action_slot(
-            slots["closeout"], resolver, registry, factory
-        )
-        # With the facade, "prefetch" records the signals for the batch
-        # connect below — same fail-fast property, now a plan message.
-        prefetch_action_signals(
-            setup_plans + per_step_plans + closeout_plans, registry, factory
-        )
-
-    detectors = _build_request_detectors(factories, devices_config, free_run=not strict)
-    movables = [build_movable(factories, target) for target in axis_resolved]
-    created.extend(factories.created)
-
-    # ---- phase 3: in-plan connects (still pre-claim) ----------------------
-    if factories.created:
-        yield from _connect_in_batches(factories.created, mock=session._mock)
-    telemetry_selected: dict[str, list[str]] = {}
-    telemetry_readables: list = []
-    if telemetry_enabled:
-        telemetry_readables, telemetry_selected = yield from _connect_telemetry_plan(
-            session, save_set, scalar_policy
-        )
-        created.extend(telemetry_readables)
-    # Telemetry is soft: appended as extra snapshot columns, never the
-    # reference (index 0 stays the save set's).
-    all_detectors = list(detectors) + telemetry_readables
-    if controller is not None and not session._mock:
-        # Fail fast on an unreachable shot-control PV (the session does this
-        # at attach time; in-plan it joins the pre-claim connect stage).
-        yield from _await_in_plan(controller.connect_setters)
-
-    if not all_detectors:
-        raise GeecsConfigurationError(
-            "ScanRequest produced no detectors (empty effective device set) — "
-            "nothing would be recorded"
-        )
-    if strict:
-        if controller is None:
-            raise GeecsConfigurationError(
-                "strict_shot_control requires a reachable shot-control device. "
-                "Use free-run mode for free-running trigger acquisition."
-            )
-        controller.require_strict_single_shot()
-
-    spec = build_step_scan_spec(
-        request,
-        axis_resolved,
-        applied_defaults=applied_defaults,
-        slots=slots,
-        dropped_unserved=dropped_unserved,
-        dropped_unserved_devices=dropped_unserved_devices,
-        disconnected_devices=disconnected_devices,
-        telemetry_selected=telemetry_selected if telemetry_enabled else {},
-        capture_devices=capture_devices,
-        native_image_save=native_image_save,
-        submission=submission,
-    )
-    if request.mode is ScanRequestMode.NOSCAN:
-        motor_arg: Any = None
-    elif len(movables) == 1:
-        motor_arg = movables[0]
-    else:
-        motor_arg = movables
 
     # ---- phase 4: the claim boundary --------------------------------------
     scan_number, scan_folder = claim_scan_number(session.experiment)
@@ -757,18 +464,7 @@ def _scan_request_body(
     inner = session.build_claimed_scan_plan(
         scan_number=scan_number,
         scan_folder=scan_folder,
-        detectors=all_detectors,
-        motor=motor_arg,
-        positions=spec.positions,
-        shots_per_step=request.capture.shots_per_step,
-        strict=strict,
-        controller=controller,
-        description=request.description,
-        md=spec.md,
-        scan_info_overrides=spec.scan_info,
-        setup=setup,
-        per_step=per_step,
-        closeout=closeout,
+        **prepared.as_claimed_kwargs(),
         # Decision-4 activation (issue #641, PR #645): under the queue a
         # failed move pauses — the RE Manager renders the paused state and
         # resume/stop are first-class queue verbs.  The headless door
